Ajustes CRUD materia

+++ src/main/java/com/Horarios/Horarios/service/MateriaService.java
@@ -1,30 +1,19 @@
 package com.Horarios.Horarios.service;
 
+import com.Horarios.Horarios.mappers.MateriaMapper;
 import com.Horarios.Horarios.model.Materia;
+import com.Horarios.Horarios.model.dto.MateriaDto;
 import com.Horarios.Horarios.repository.IMateriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class MateriaService {
 
     @Autowired
     private IMateriaRepository iMateriaRepository;
-
-    public Map<String, Object> guardarMateria(Materia materia){
-        Map<String, Object> map = new HashMap<String,Object>();
-        if(!consultarPorNombre(materia.getNombre()).isPresent() && validarObjeto(materia.getNombre())){
-            map.put("response", iMateriaRepository.save(materia).getId());
-            return map;
-        }else{
-            return map;
-        }
-    }
 
     public boolean validarObjeto(String nombreMateria){
         if(nombreMateria == null){
@@ -36,29 +25,66 @@ public class MateriaService {
         return true;
     }
 
-    public List<Materia> consultarMaterias(){
-        return iMateriaRepository.findAll();
-    }
-
-    public void eliminarMateria(Long id){
-
-        iMateriaRepository.deleteById(id);
-    }
-
-    public Materia editarMateria(Materia materia){
-        Materia validacionMateria = null;
-        if(materia.getId() != null && iMateriaRepository.existsById(materia.getId())){
-            validacionMateria = iMateriaRepository.save(materia);
+    public Map<String, Object> guardarMateria(Materia materia){
+        Map<String, Object> map = new HashMap<String,Object>();
+        if(consultarPorNombre(materia.getNombre()) == null && validarObjeto(materia.getNombre())){
+            map.put("response", iMateriaRepository.save(materia).getId());
+            return map;
+        }else{
+            map.put("error", null);
+            return map;
         }
-        return validacionMateria;
     }
 
-    public Optional<Materia> consultarPorId(Long id){
-        return iMateriaRepository.findById(id);
+    public List<MateriaDto> consultarMaterias(){
+        return MateriaMapper.convertirEntitysADtos(iMateriaRepository.findAll());
     }
 
-    public Optional<Materia> consultarPorNombre(String nombre){
-        return iMateriaRepository.findByNombre(nombre);
+    public Map<String, Object> eliminarMateria(Long id){
+
+        Map<String, Object> map = new HashMap<>();
+        if(iMateriaRepository.existsById(id)){
+            map.put("response", id);
+            iMateriaRepository.deleteById(id);
+            return map;
+        }else {
+            map.put("error", null);
+            return map;
+        }
+    }
+
+    public Map<String, Object> editarMateria(Materia materia){
+        Map<String, Object> map = new HashMap<>();
+        if(materia.getId() != null && iMateriaRepository.existsById(materia.getId())){
+            map.put("response", iMateriaRepository.save(materia).getId());
+            return map;
+        }else{
+            map.put("error", null);
+            return map;
+        }
+
+    }
+
+    public MateriaDto consultarPorId(Long id){
+        Optional<Materia> optional = iMateriaRepository.findById(id);
+        if(optional.isPresent()){
+            return MateriaMapper.convertirEntityADto(optional.get());
+        }else{
+            return null;
+        }
+
+    }
+
+    public MateriaDto consultarPorNombre(String nombre){
+
+        Optional<Materia> optional = iMateriaRepository.findByNombre(nombre);
+        if (optional.isPresent()){
+            return MateriaMapper.convertirEntityADto(optional.get());
+
+        }else{
+            return null;
+        }
+
     }
 }
